Project5/requirements.txt
pymongo>=4.6
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
tqdm>=4.66
ip2location>=8.10
